# internal/repositories/postgres/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.13.0

package postgres

import (
	"database/sql"
	"time"
)

type Blog struct {
	ID          int32
	Name        sql.NullString
	Description sql.NullString
}

type Content struct {
	ID          int32
	UserEmail   sql.NullString
	Name        sql.NullString
	Description sql.NullString
}

type Session struct {
	ID           int32
	UserEmail    sql.NullString
	RefreshToken sql.NullString
	AccessToken  sql.NullString
	UserAgent    sql.NullString
	Ip           sql.NullString
	ExpiresIn    int64
	CreatedAt    time.Time
}

type User struct {
	ID        int32
	Name      sql.NullString
	Password  string
	Email     string
	Role      string
	IsDeleted sql.NullBool
	Verifay   bool
}